FE: split column definition helper to separate files
